Consolidate intersection headers
