+++ cpp/include/cuspatial/intersection.cuh
