+++ cpp/include/cuspatial/intersection.hpp
