+++ cpp/tests/spatial/intersection/linestring_intersection_test.cu
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, NVIDIA CORPORATION.
+ * Copyright (c) 2022-2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,8 +21,8 @@
 
 #include <cuspatial/error.hpp>
 #include <cuspatial/geometry/vec_2d.hpp>
+#include <cuspatial/intersection.cuh>
 #include <cuspatial/iterator_factory.cuh>
-#include <cuspatial/linestring_intersection.cuh>
 #include <cuspatial/traits.hpp>
 
 #include <cudf/column/column.hpp>
